pyproject.toml
[project]
name = "github-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.10.6",
    "google-genai>=1.28.0",
    "google-generativeai>=0.8.5",
    "pygithub>=2.7.0",
    "rich>=14.1.0",
    "ruff>=0.12.7",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]
